backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
tensorflow>=2.13.0
opencv-python
Pillow
numpy
pydantic
firebase-admin
kagglehub
scipy

# Training dependencies
matplotlib>=3.5.0
seaborn>=0.12.0
scikit-learn>=1.2.0

# AI/LLM
google-generativeai

# Testing
pytest
pytest-asyncio
httpx
